Move ert_data to ert.data
