+++ .github/workflows/coverage.yml
@@ -46,7 +46,7 @@ jobs:
         # of 3.0.0) will create an incorrect report when a directory matches
         # the package name
         cd tests
-        pytest ert_tests --cov=ert --cov=ert3 --cov=ert_data --cov=ert_shared --cov=ert_logging --cov=res --cov-report=xml:cov.xml
+        pytest ert_tests --cov=ert --cov=ert3 --cov=ert_shared --cov=ert_logging --cov=res --cov-report=xml:cov.xml
 
     - name: Run res tests
       if: matrix.system-under-test == 'res'
