+++ .mypy-strict.ini
@@ -8,7 +8,7 @@ ignore_errors = True
 [mypy-ert.gui.*]
 ignore_errors = True
 
-[mypy-ert_data.*]
+[mypy-ert.data.*]
 ignore_errors = True
 
 [mypy-cloudpickle.*]
