+++ .mypy.ini
@@ -93,11 +93,11 @@ ignore_errors = True
 ignore_missing_imports = True
 ignore_errors = True
 
-[mypy-ert_data.*]
+[mypy-ert.data.*]
 ignore_missing_imports = True
 ignore_errors = True
 
-[mypy-ert_data.measured]
+[mypy-ert.data.measured]
 ignore_missing_imports = False
 ignore_errors = False
