+++ src/ert/__init__.py
@@ -1,3 +1,3 @@
-from ert_data import MeasuredData
+from .data import MeasuredData
 
 __all__ = ["MeasuredData"]  # Used by semeio
