--- src/ert_data/__init__.py
@@ -1,3 +1,0 @@
-from .measured import MeasuredData
-
-__all__ = ["MeasuredData"]
