+++ src/ert_shared/storage/extraction.py
@@ -1,7 +1,7 @@
 from collections import defaultdict
 from typing import Dict, List, Mapping, Optional, Union, TYPE_CHECKING
 
-from ert_data.measured import MeasuredData
+from ert.data import MeasuredData
 from res.enkf.enums.enkf_obs_impl_type_enum import EnkfObservationImplementationType
 from ert_shared.feature_toggling import feature_enabled
 from ert_shared.services import Storage
